+++ NEWS.md
@@ -1,4 +1,5 @@
 # CFAEpiNow2Pipeline (development version)
 
+* Azure Blob file download utilities
 * CI running on Ubuntu only & working pkgdown deploy to Github Pages
 * Initial R package with checks running in CI